Make the tests work with JavaConfig. Also add Eclipse launcher
