--- src/test/resources/graffitab-servlet-test.xml
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans:beans xmlns="http://www.springframework.org/schema/mvc"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-             xmlns:beans="http://www.springframework.org/schema/beans"
-             xmlns:mvc="http://www.springframework.org/schema/mvc"
-             xmlns:context="http://www.springframework.org/schema/context"
-             xsi:schemaLocation="http://www.springframework.org/schema/mvc http://www.springframework.org/schema/mvc/spring-mvc-3.2.xsd
-
-        http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
-        http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-3.1.xsd">
-
-    <beans:bean id="delegateJacksonHttpMessageConverter" class="org.springframework.http.converter.json.MappingJackson2HttpMessageConverter"/>
-    <beans:bean id="jacksonHttpMessageConverter" class="com.graffitab.server.config.spring.CustomMappingJackson2HttpMessageConverter"/>
-
-    <annotation-driven>
-        <mvc:message-converters>
-            <beans:ref bean="jacksonHttpMessageConverter"/>
-        </mvc:message-converters>
-        <mvc:argument-resolvers>
-            <beans:bean class="com.graffitab.server.config.spring.JsonDtoArgumentResolver" />
-        </mvc:argument-resolvers>
-    </annotation-driven>
-
-    <!-- Controllers -->
-    <context:component-scan base-package="com.graffitab.server.api" use-default-filters="false">
-        <context:include-filter type="annotation" expression="org.springframework.stereotype.Controller"/>
-        <context:include-filter type="annotation" expression="org.springframework.stereotype.Component"/>
-    </context:component-scan>
-
-</beans:beans>
