--- src/test/resources/spring-context-test.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:context="http://www.springframework.org/schema/context"
-    xmlns:task="http://www.springframework.org/schema/task"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans
-               			http://www.springframework.org/schema/beans/spring-beans-3.2.xsd
-               			http://www.springframework.org/schema/context
-               			http://www.springframework.org/schema/context/spring-context-3.1.xsd
-            			http://www.springframework.org/schema/task
-            			http://www.springframework.org/schema/task/spring-task-3.0.xsd">
-
-    <context:annotation-config/>
-
-    <context:component-scan base-package="com.graffitab.server" />
-    
-    <bean id="multipartResolver" class="org.springframework.web.multipart.commons.CommonsMultipartResolver">
-    </bean>
-
-    <bean id="messageSource"
-      class="org.springframework.context.support.ResourceBundleMessageSource">
-  		<property name="basenames">
-		    <list>
-		      <value>messages</value>
-		    </list>
-  		</property>
-	</bean>
-
-     <!--  Async and scheduling -->
-    <task:annotation-driven />
-      
-    <bean id="restTemplate" class="org.springframework.web.client.RestTemplate" />
-    <bean id="jacksonHttpMessageConverter" class="org.springframework.http.converter.json.MappingJackson2HttpMessageConverter"/>
-    
-    <import resource="classpath:jdbc-test.xml"/>
-    <import resource="classpath:spring/persistence.xml"/>
-
-</beans>
